mm-video: vidc: Parse Seq start code and Entry point code for VC1
With out this change, the decoder test app parser was sending the
sequence start code and entry point start code at the end of the
current frame. Hence there was no reconfig message coming from
the video core. This change will prepend the seq start code and
entry point start code data at the start of the next frame, as
expected by the video core.

This change is required in the decoder test app parser to
support reconfig data in vc1 bitstream.

Change-Id: I579496cdbadd6e37fc57128c6f2911dddaccab60
CRs-fixed: 342166