vidc: vdec: Do not flush frame parser if only SPS/PPS is queued.
If seek/flush occurs after sps/pps buffers given to video component,
the frameparser will be flushed and hence sps/pps are not sent to
the driver which is causing the decoding failures of subsequent
frames. This commit avoids the frame parser flush if codec header
buffers are not sent to the driver yet.

Change-Id: I2e75dd6b8fde0cf5f5cf8bf810f95a98718b1c65
CRs-fixed: 353830