mm-video: vidc: Update TS entry removal logic for interlaced clips
Some of the interlaced clips have non-duplicate time stamps for
different fields/ETBs corresponding to a frame. Because of this,
time stamp parser table is getting full due to a problem in TS
entry removal. This commit takes care of this corner case and
removes next TS entry if a duplicate entry is not found for an
interlaced clip.

Change-Id: I738c411d5b2e377c3bf9cf0809e51b3a2878a9fa
CRs-fixed: 356619